+++ jirust-cli/src/executors/jira_commands_executors.rs
@@ -1,7 +1,13 @@
+pub mod jira_project_executor;
 pub mod jira_version_executor;
 
 /// Trait to be impelemented to execute Jira commands in Jira commands executors
 pub trait ExecJiraCommand {
+    /// Asynchronous function to execute Jira command.
+    ///
+    /// # Returns
+    ///
+    /// * A Result with a unit type or a Box with a dyn std::error::Error trait.
     fn exec_jira_command(
         &self,
     ) -> impl std::future::Future<Output = Result<(), Box<dyn std::error::Error>>> + Send;