+++ jirust-cli/src/lib.rs
@@ -7,6 +7,7 @@ use crate::executors::jira_commands_executors::jira_version_executor::VersionExe
 use clap::Parser;
 use config::config_file::ConfigFile;
 use executors::config_executor::ConfigExecutor;
+use executors::jira_commands_executors::jira_project_executor::ProjectExecutor;
 use executors::jira_commands_executors::ExecJiraCommand;
 use std::env::Args;
 use std::io::{Error, ErrorKind};
@@ -15,6 +16,7 @@ pub mod args;
 pub mod config;
 pub mod executors;
 pub mod runners;
+pub mod utils;
 
 /// Manages the loading of the CLI configuration
 ///
@@ -127,6 +129,10 @@ pub async fn process_command(
         Commands::Version(args) => {
             let version_executor = VersionExecutor::new(cfg_data, args.version_act, args);
             version_executor.exec_jira_command().await?
+        }
+        Commands::Project(args) => {
+            let project_executor = ProjectExecutor::new(cfg_data, args.project_act, args);
+            project_executor.exec_jira_command().await?
         }
     }
     Ok(())