+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 ## [Unreleased]
+### Added
+- Added Jira projects-related commands
+
+### Changed
+- Improved docs and tests
 
 ## [0.7.4] 2024-08-14
 ### Changed